Adds scalatest tests
